+++ packages/koot/utils/init-node-env.js
@@ -93,6 +93,9 @@ module.exports = () => {
         // 当前是否是测试模式
         KOOT_TEST_MODE: JSON.stringify(false),
 
+        // 当前是否是内部开发模式
+        KOOT_DEVELOPMENT_MODE: JSON.stringify(false),
+
         // 开发环境启动时间
         // KOOT_DEV_START_TIME: Date.now(),